feat (login) : add feature to read key & value in application.properties
