+++ src/main/java/atdd/path/jwt/JwtTokenProvider.java
@@ -1,5 +1,7 @@
 package atdd.path.jwt;
 
+import lombok.Value;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 @Component
