+++ src/main/java/atdd/path/jwt/Properties.java
@@ -1,0 +1,24 @@
+package atdd.path.jwt;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@EnableConfigurationProperties(ReadProperties.class)
+public class Properties {
+    private String secretKey;
+    private int expireLength;
+
+    public Properties(ReadProperties readProperties) {
+        this.secretKey = readProperties.getSecretKey();
+        this.expireLength = readProperties.getExpireLength();
+    }
+
+    public String getSecretKey() {
+        return secretKey;
+    }
+
+    public int getExpireLength() {
+        return expireLength;
+    }
+}
