--- src/test/java/atdd/path/jwt/Properties.java
@@ -1,14 +1,0 @@
-package atdd.path.jwt;
-
-public class Properties {
-    private String secretKey;
-    private int expireLength;
-
-    public String getSecretKey() {
-        return secretKey;
-    }
-
-    public int getExpireLength() {
-        return expireLength;
-    }
-}
